fix(Placeholder): fix flash when url bar resizes
When scrolling on an ios device, the browser bar may
shrink or grow which resizes the page in height. This
forces the elements to recalculate its size which caused
a flash. This removes the height from triggering recalculating.
It would be better to indicate when the user is scrolling and
only recalculate when its save to do so.