feat(*): improve performance
This commit changes the whole implementation of when
the component updates. It will now only update
on scroll and resize events and will then calculate
the bounding rects of the elements. The change was required
because listening to bounding client rects in an request animation
frame is expensive. Also some components were rerendered multiple
times which was unnecessary.
Further some smart should update hooks will now prevent useless
rerenderings.