fix(Sticky): remove gab between component and StickyScrollUp
When scrolling fast and the page is not updated fast enought
the sticky offset is calculated before the position fixed
is set for the StickyScrollUp component. In this situation
a gap between the elements will appear. By setting the offset
the the height (which is always the highest value possible)
prevents this to happening.