feat(Sticky): add overflowScroll property
The overflowScroll property allows to define how the component
should behave in case its bigger than the viewport.
The default behaviour is 'end' which only scrolls when the bottom
of the container is reached. This is the same as it was before.
A new option 'flow' was introduced to make the content more accessible,
it scrolls with the flow of the scroll direction, therefore the content
is better accessible in most situations.