feat(StickyScrollUp): calculate start position by default
When the StickyScrollUp component is not the first
element on the page but will scroll in later it the
component needs to be aware where it starts. This was
solved with the defaultOffsetTop which was only a fast
and error prone way to set this value from the outside.
With this commit, it will calculate its start position
on it's own so there is no need for the defaultOffsetTop
property anymore. Therefore its deprecated.

Issue: #11